Project Pizza GUI/requirements.txt
Pillow = 10.2.0 
#Installation Instructions
    # go on bash/command line
    # pip install pillow